chore: refactor layered state adaptor into separate files
